### 2023-07-27 #5696 @kyleniemeyer
@openjournals/dev I can see that the repository has a GPL 3 license, any idea why editorialbot didn't find it?
