### 2023-07-27 #5696 @xuanxu
> @openjournals/dev I can see that the repository has a GPL 3 license, any idea why editorialbot didn't find it?

editorialbot couldn't find it because the license is not present in the paper branch
