### 2023-07-31 #5696 @RainierBarrett
Yes, I'd be happy to.
