### 2023-08-01 #5715 @editorialbot
[REVIEW]: PLIP: A hybrid python/C++ package for linearized machine learning interatomic potentials
**Submitting author:** @aksam432 (<a href="http://orcid.org/0000-0002-9880-4482">Akshay Krishna Ammothum Kandy</a>)
**Repository:** https://github.com/LAM-GROUP/PLIP
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.4
