### 2023-08-03 LAM-GROUP/PLIP#4 @RainierBarrett
JOSS Paper equations are missing some definitions
Hello! I've been tapped as a reviewer for your recent submission to JOSS!
I've noticed that the PDF proof linked [at the JOSS review issue thread](https://github.com/openjournals/joss-reviews/issues/5715) doesn't contain definitions of all the symbols in the equations at the top of page 2:
![image](https://github.com/LAM-GROUP/PLIP/assets/13456066/374cadf4-6209-4e57-a829-531e6d4c0ad8)

I think the paper should include definitions of all the symbols used in these equations, as you have [in your docs](https://plip.readthedocs.io/en/latest/source/model.html). That is, it's not clear what $f_c$, $f_s$, or $f_n$ are in the paper alone.
