### 2023-08-03 LAM-GROUP/PLIP#5 @RainierBarrett
Please add community guidelines
In connection with your [ongoing JOSS review](https://github.com/openjournals/joss-reviews/issues/5715), please add a section/subsection to your docs regarding community guidelines (contributing, reporting issues, and seeking support).
