### 2023-08-17 #5715 @Materials-Informatics-Laboratory
@aksam432 quick question regarding the paper. When you say that your method represents an "interpretable MLIP", what exactly does interpretable mean in this context? The only portion I could find that might fall into this category is the use of LASSO, but I'm not sure if I would go so far as to say it gives you an "interpretable" MLIP.
